Give the piped exec composition one home
The state-aware `Piped` exec and the one-shot streaming backend need the same
composition: acquire an apartment reference the escaping object can hold, start
the process, and expose its three pipe handles alongside the closures that wait
on it, close its stdin, and terminate it.

Extracted from the state-aware arm unchanged, so the second caller reuses it
rather than deriving it again.

Copilot-Session: 82dd9a10-d9ab-4a9c-b68a-c585d07fc959
